chore: cleans up infra tasks
