+++ e2e/infra/cluster_operations.go
@@ -7,22 +7,6 @@ import (
 	"github.com/maistra/istio-workspace/test/shell"
 )
 
-// CreateNewApp creates new project with a given name, deploys simple datawire/hello-world app and exposes route to
-// it service
-func CreateNewApp(name string) {
-	<-shell.Execute(NewProjectCmd(name)).Done()
-
-	UpdateSecurityConstraintsFor(name)
-
-	<-shell.ExecuteInDir(".",
-		"oc", "new-app",
-		"--docker-image", "datawire/hello-world",
-		"--name", name,
-		"--allow-missing-images",
-	).Done()
-	shell.ExecuteAll("oc expose svc/"+name, "oc status")
-}
-
 // UpdateSecurityConstraintsFor applies anyuid and privileged constraints to a given namespace
 func UpdateSecurityConstraintsFor(namespace string) {
 	LoginAsTestPowerUser()
