+++ e2e/infra/deployment.go
@@ -2,6 +2,7 @@ package infra
 
 import (
 	"fmt"
+	"github.com/maistra/istio-workspace/test/shell"
 	"os"
 
 	"github.com/onsi/gomega"
@@ -32,4 +33,25 @@ func DeleteFile(filePath string) {
 
 func NewProjectCmd(name string) string {
 	return fmt.Sprintf(`oc new-project %s --description "%s"`, name, "istio-workspace test project")
+}
+
+// CreateNewApp creates new project with a given name, deploys simple datawire/hello-world app and exposes route to
+// it service
+func CreateNewApp(name string) {
+	<-shell.Execute(NewProjectCmd(name)).Done()
+
+	UpdateSecurityConstraintsFor(name)
+
+	<-shell.ExecuteInDir(".",
+		"oc", "new-app",
+		"--docker-image", "datawire/hello-world",
+		"--name", name,
+		"--allow-missing-images",
+	).Done()
+	shell.ExecuteAll("oc expose svc/"+name, "oc status")
+}
+
+
+func DeployHelloWorldCmd(name string) string {
+	return "oc new-app --docker-image datawire/hello-world --name " + name + " --allow-missing-images"
 }
